Add Gradient to CircularProgressBar (#89)
